# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ReliefMesh API
  version: 0.1.0
  summary: Offline-first, privacy-preserving coordination for local disaster-response exercises and community aid.
  description: |
    REST API of a self-hosted ReliefMesh instance.

    **ReliefMesh is not an emergency service.** For immediate danger to life, health, fire, crime or other
    emergencies, contact the responsible local emergency services. Nothing in this API notifies authorities,
    dispatches responders or performs medical triage.

    ## Authentication
    Sessions use an HTTP-only, `SameSite=Strict` cookie (`__Host-rm_session` over HTTPS, `rm_session` in local
    development). Every state-changing request from an authenticated session must send the CSRF token returned by
    `POST /auth/login` or `GET /auth/session` in the `X-CSRF-Token` header. Browsers must also send a matching
    `Origin`.

    ## Authorization
    All authorization is enforced on the server. Records the caller may not see are reported as `404` so their
    existence is not revealed. Every view contains a `permissions` object computed on the server.

    ## Privacy
    Protected values (contact details and exact locations) are never part of regular responses. They are only
    returned by the `.../protected` endpoints, which record an audit event that is visible to the data subject.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Same origin as the web app
...
security:
  - session: []
paths:
...
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List local accounts
      responses:
        "200": { description: Users, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/User" } } } } } } }
    post:
      tags: [admin]
      summary: Create an account with a one-time temporary password
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, display_name, roles]
              properties:
                username: { type: string, pattern: "^[a-z0-9][a-z0-9._@\\-]{2,63}$" }
                display_name: { type: string, maxLength: 80 }
                roles: { type: array, items: { $ref: "#/components/schemas/Role" }, minItems: 1 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  temporary_password: { type: string }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    patch:
      tags: [admin]
      summary: Change display name, roles, active state or unlock
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                display_name: { type: string }
                roles: { type: array, items: { $ref: "#/components/schemas/Role" } }
                is_active: { type: boolean }
                unlock: { type: boolean }
      responses:
        "200": { description: User, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{id}/reset-password:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [admin]
      summary: Reset a password (no e-mail involved); ends all sessions of the user
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "200": { description: Temporary password, content: { application/json: { schema: { type: object, properties: { temporary_password: { type: string } } } } } }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-rm_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    CSRF: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [validation_failed, forbidden, not_found, invalid_transition, over_allocation, version_conflict, csrf_invalid] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            details: { type: object }
            request_id: { type: string }
    Role: { type: string, enum: [requester, volunteer, coordinator, organization_manager, admin] }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
        roles: { type: array, items: { $ref: "#/components/schemas/Role" } }
        is_active: { type: boolean }
        must_change_password: { type: boolean }
        locked: { type: boolean }
        availability: { type: string }
        last_login_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
...